+++ tests/helpers/TokenizerTest.php
